theme multiplication is finished !

+++ tsvp_posvt/include/iobigint.h
@@ -4,7 +4,7 @@
 #include <stddef.h>
 #include <stdint.h>
 
-int get_bigint(unsigned short*, size_t, int*);
-void print_bigint(const unsigned short*, size_t, int);
+int get_bigint(unsigned*, size_t, int*);
+void print_bigint(const unsigned*, size_t, int);
 
 #endif //__IOBIGINT_H__
